requirements.txt
discord.py
discord-py-interactions
structlog
python-dotenv
requests
textdistance
arrow
geopy
progressbar